refactor: constant to enum class
